fix several errors on fetchrange3 script

+++ bin/fetchrange3.pl
@@ -31,7 +31,6 @@ use threads::shared;
 use Thread::Semaphore;
 use Storable;
 use Carp 'verbose';
-
 use Ham::APRS::IS;
 use Ham::APRS::FAP qw(parseaprs);
 use DBI;
@@ -40,19 +39,27 @@ use Math::Trig qw(:great_circle deg2rad rad2deg);
 use FindBin;
 use lib $FindBin::Bin;
 use Geo::Coordinates::UTM;
-
+use DateTime;
 use Socket qw(SOL_SOCKET SO_RCVBUF);
-
+use IO::Socket::INET;
+use JSON;
+use Config;
+# --------------------- #
 my $id = 'GLDDBF61';
 my %countries :shared;
 
 
 use LatLngCoarse2;
+#
+## set prt =1 if you want printing debugging information
+#
 
+my $prt=0;
+my $pgmversion = "0.4.0";
 
-use JSON;
-
+#
 # load the configuration file
+#
 my $filename = "../config/binconfig.json";
 
 my $json_text = do {
@@ -70,13 +77,20 @@ use Socket;
 my $address = inet_ntoa(
         	scalar gethostbyname( $host || 'localhost' )
     );
-warn "\n === $host === \n";
-warn "\n === $hostname === \n";
-warn "\n === $address === \n";
+warn "\n OGN Range analyzer Version: $pgmversion \n";
+warn "Host === $host === \n";
+warn "Host === $hostname === \n";
+warn "IP addr === $address === \n";
+warn "OSname: $Config{osname}\n";
+warn "OSname: $Config{archname}\n";
+#warn "ProcessID $PID\n";
+#
+#Get the credential from the config file
+#
 my $json = JSON->new;
 my $data = $json->decode($json_text);
 
-my $db_dsn = $data->{config}->{dsn} || die "no dsn specified";
+my $db_dsn      = $data->{config}->{dsn} || die "no dsn specified";
 my $db_username = $data->{config}->{dbusername} || die "no user specified";
 my $db_password = $data->{config}->{dbpassword} || die "no user specified";
 
@@ -136,7 +150,17 @@ while(1) {
     $ready_semaphore->down(100);
     
     my $avail = threads->create( \&handleAvailablity );
-    foreach my $server ( @servers ) {
+    foreach my $server ( @servers ) {	
+	my $name="localhost";
+	if ($server =~ /glider*/ ) {
+		$name=substr($server, 0, 22);
+	}
+	if ($server =~ /aprs*/ )
+	{
+		$name=substr($server, 0, 18);
+	}
+	$address = inet_ntoa(inet_aton($name));
+	print "\n<<<< Connecting Server: $server  at IP addr: $address <<<<\n";
 	threads->create( \&handleServer, $server );
     }
     
@@ -175,9 +199,18 @@ sub handleServer {
     my $sth_addglider =  $db->prepare( 'insert into gliders ( callsign ) values ( ? )' );
     my $sth_history =  $db->prepare( 'insert into history values ( now(), ?, ?, ? )' );
 
-    print "connecting to server $server\n";
-    my $is = new Ham::APRS::IS( $server, 'OGR', 'appid' => 'ognrange.glidernet.org 0.4.0');
-    
+    my $name   =  "localhost";
+    if ($server =~ /glider*/ )
+	{
+		$name=substr($server, 0, 22);
+	}
+    if ($server =~ /aprs*/)
+	{
+		$name=substr($server, 0, 18);
+	}
+    my $address = inet_ntoa(inet_aton($name));
+    if ($prt) {print "connecting to server $server at IP addr: $address \n";}
+    my $is = new Ham::APRS::IS( $server, 'OGNRANGE', 'appid' => 'ognrange.glidernet.org 0.4.0');
 
     open( OUT, ">>", "/dev/null" );
 #    open( OUT, ">>", $server );
@@ -189,43 +222,46 @@ sub handleServer {
     while(1) {
 	$is->connect('retryuntil' => 10) || print "Failed to connect: $is->{error}";
 
-	$is->sock()->setsockopt(SOL_SOCKET,SO_RCVBUF,256*1024);
+	if   ($is->connected()) {$is->sock()->setsockopt(SOL_SOCKET,SO_RCVBUF,256*1024);}
 	
 	while($is->connected()) {
 	    	    
 	    # make sure we send a keep alive every 90 seconds or so
 	    my $now = time();
 	    if( $now - $lastkeepalive > 60 ) {
-		$is->sendline('# ognrange.glidernet.org 212.13.204.217');
+		$is->sendline('# ognrange.glidernet.org 51.254.32.187 ');
 		$lastkeepalive = $now;
 	    }
 	    
 	    # check to make sure we emit the stations at least once a day
 	    if( date($now) ne $today ) { 
 		$today = date($now);
-		%stations_loc = (); %stations_ver = ();
-		print "resetting stations for change of date";
-		print OUT "\n--------------- $today ---------------\n";
-		warn      "\n--------------- $today ---------------\n";
-		warn      "\n--------------- $now   ---------------\n";
+		if ($full) {
+			%stations_loc = (); %stations_ver = ();
+			if ($prt) {print "resetting stations for change of date";}
+			print OUT "\n-------New date: -------- $today ---------------\n";
+			warn      "\n-------New date: -------- $today ---------------\n";
+			warn      "\n-------New date: -------- $now   ---------------\n";
 
 		# hide old stations so they don't linger forever
-		$db->do( "create temporary table z as select sl.station, datediff(now(),max(sl.time)) a, least(greatest(5,count(sl.time)*2),21) b from stationlocation sl group by 1 having a > b" );
-		$db->do( "update stations set active='N' where id = (select station from z where z.station = stations.id)" );
-		$db->do( "drop temporary table z" );
+			$db->do( "create temporary table z as select sl.station, datediff(now(),max(sl.time)) a, least(greatest(5,count(sl.time)*2),21) b from stationlocation sl group by 1 having a > b" );
+			$db->do( "update stations set active='N' where id = (select station from z where z.station = stations.id)" );
+			$db->do( "drop temporary table z" );
 
 		# ensure we have up to date coverage map once a day
-		$db->do ( "truncate estimatedcoverage " );
-		$db->do ( "insert into estimatedcoverage select station, ref, avg(strength) s,sum(count) c from positions_mgrs p group by station, ref having (s > 75 and c > 20) or s > 105" );
-		$db->do ( "truncate roughcoverage " );
-		$db->do ( "insert into roughcoverage select station, concat(left(ref,6),mid(ref,8,1)) r, avg(strength) s,sum(count) c from positions_mgrs p group by station, r " );
-		warn      "\n--------------- $now   ---------------\n";
+			$db->do ( "truncate estimatedcoverage " );
+			$db->do ( "insert into estimatedcoverage select station, ref, avg(strength) s,sum(count) c from positions_mgrs p group by station, ref having (s > 75 and c > 20) or s > 105" );
+			$db->do ( "truncate roughcoverage " );
+			$db->do ( "insert into roughcoverage select station, concat(left(ref,6),mid(ref,8,1)) r, avg(strength) s,sum(count) c from positions_mgrs p group by station, r " );
+	    		my $now = time();
+			warn      "\n--------------- $now   ---------------\n";
+		}
 	    }
 
 	    
 	    my $l = $is->getline(120);
 	    if (!defined $l) {
-		print "\n".date($now).": $server: failed getline: ".$is->{'error'}."\n";
+		if ($prt) {print "\n".date($now).": $server: failed getline: ".$is->{'error'}."\n";}
 		$is->disconnect();
 		last;
 	    }
@@ -234,13 +270,13 @@ sub handleServer {
 
 	    
 	    if( $l =~ /^\s*#/ || $l eq '' ) {
-#		print "\n$l\n";
+#		if ($prt) {print "\n$l\n";}
 		next;
 	    }
 
 	    if( $full ) {
 		$l =~ /^dup\s+(.*)$/;
-		print "\n->".$1."<-\n";
+		if ($prt) {print "\n->".$1."<-\n";}
 		$l = $1;
 	    }
 
@@ -289,10 +325,10 @@ sub handleServer {
 				    my $reduced = ($1||'0'); $reduced .= $2.$3;
 
 				    # and store the record in the db
-				    $sth_mgrs->execute( $s_id, $reduced, $strength, $height, $height ) or die "Can't execute statement: $DBI::errstr";;
+				    $sth_mgrs->execute( $s_id, $reduced, $strength, $height, $height ) or warn "Can't execute statement: $DBI::errstr";;
 
 				    if( $full ) {
-					print "DUP: $s_callsign $s_id ($v1) lt='$lt_r' and lg='$lg_r'\n";
+					if ($prt) {print "DUP: $s_callsign $s_id ($v1) lt='$lt_r' and lg='$lg_r'\n";}
 				    }
 				    
 				    {
@@ -313,7 +349,7 @@ sub handleServer {
 					}
 
 					if( ($stations_status{$v1}||'U') eq 'N' ) {
-					    print "packet from $v1 which is currently shown as noppm\n";
+					    if ($prt) {print "packet from $v1 which is currently shown as noppm\n";}
 					    $stations_status{$v1} = 'U';
 					    $stations_ppm{$v1} = 0;
 					}
@@ -330,7 +366,7 @@ sub handleServer {
 				    }
 				}
 				else {
-				    print "*";
+				    if ($prt) {print "*";}
 				    {
 					lock( %stations_last );
 					$stations_last{$v1} = $now;
@@ -342,7 +378,7 @@ sub handleServer {
 			}
 			elsif( $v1 eq 'qAC' ) {
 
-			    print "--------------------------------------------------------\n";
+			    if ($prt) {print "-------------------- qAC ------------------------------------\n";}
 
 			    # qAC seems to be the beacons
 			    my $s_id = getStation( $sth_addstation, $sth_history, $sth_updatestation, $callsign );
@@ -350,17 +386,17 @@ sub handleServer {
 			    if( $packetdata{type} eq 'location' && 
 				$packetdata{latitude} && $packetdata{longitude} ) {
 
-				print "location beacon $l\n";
+				if ($prt) {print "location beacon $l\n";}
 
 				processStationLocationBeacon( $db, $callsign, $s_id, $packetdata{latitude}, $packetdata{longitude}, $packetdata{altitude}, $packetdata{comment}||'' );
 
 				if( $packetdata{comment} && $packetdata{comment} =~ /v0.2.[0-5]/ ) {
-				    print "OLD details beacon $///$packetdata{comment}l\n";
+				    if ($prt) {print "OLD details beacon $///$packetdata{comment}l\n";}
 				    processStationDetailsBeacon( $db, $callsign, $s_id, $packetdata{comment});
 				}
 			    }
 			    else {
-				print "details beacon $l\n";
+				if ($prt) {print "details beacon $l\n";}
 				processStationDetailsBeacon( $db, $callsign, $s_id, $l );
 			    }
 			}
@@ -368,18 +404,18 @@ sub handleServer {
 			    #		    print "status: $l\n";
 			}
 			else {
-			    print "\n--- new packet ---\n$l\n";
+			    if ($prt) {print "\n--- new packet ---\n$l\n";}
 			    while (my ($key, $value) = each(%packetdata)) {
-				print "$key: $value\n";
+				if ($prt) {print "$key: $value\n";}
 			    }
 			}
 		    }
 		}
 	    }
 	    else {
-		print "\n$server: --- bad packet --> \"$l\"\n";
+		if ($prt) {print "\n$server: --- bad packet --> \"$l\"\n";}
 		if ($l =~ 'RND') {
-			print "Ignore RND pacakets"
+			if ($prt) {print "Ignore RND pacakets"}
 		}
 		else {
 			warn "Parsing failed: $packetdata{resultmsg} ($packetdata{resultcode}) <==== \n";
@@ -387,7 +423,22 @@ sub handleServer {
 		}
 	    }
 	}
-	print "\nreconnecting\n";
+	my $sock   = $is->sock();
+	my $nownow = time();
+	my $dt     = DateTime->now;
+	my $dmy    = $dt->dmy;
+	my $hms    = $dt->hms;
+	my $name   =  "localhost";
+	if ($server =~ /glider*/ )
+	{
+		$name=substr($server, 0, 22);
+	}
+	if ($server =~ /aprs*/)
+	{
+		$name=substr($server, 0, 18);
+	}
+	$address = inet_ntoa(inet_aton($name));
+	print "\n>>>> reconnecting $server Server: $name is at IP: $address Date and Time: $nownow  $dmy $hms <<<<\n";
 	$is->disconnect() || print "Failed to disconnect: $is->{error}";
 	sleep(30);
     }
@@ -400,7 +451,7 @@ sub getStation {
 	    return 0;
     }
 
-    if ($station =~ /FNB*/ || $station =~ /XCG*/ ) {
+    if ($station =~ /FNB*/ || $station =~ /XCG*/ ||  $station =~ /OGN*/ || $station =~ /RELAY*/) {
 	    return 0;
     }
 
@@ -413,13 +464,13 @@ sub getStation {
 	    $s_id = $sth_add->{mysql_insertid};
 	    $station_id{lc $station} = $s_id;
 	    $station_name{ lc $station } = $station;
-	    print "\nnew station $station => $s_id\n";
-	    $sth_history->execute( $s_id, 'new', "New station $station" ) or die "Can't execute statement: $DBI::errstr";;
+	    if ($prt) {print "\nnew station $station => $s_id\n";}
+	    $sth_history->execute( $s_id, 'new', "New station $station" ) or warn "Can't execute statement: $DBI::errstr";;
 	}
 	elsif( $sth_supdate && ($station_name{ lc $station }||$station) ne $station ) {
-	    $sth_supdate->execute( $station, $s_id ) or die "Can't execute statement: $DBI::errstr";;	
-	    print "\nrenamed station $station => $s_id\n";
-	    $sth_history->execute( $s_id, 'renamed', $station_name{ lc $station } . " now $station" ) or die "Can't execute statement: $DBI::errstr";;
+	    $sth_supdate->execute( $station, $s_id ) or warn "Can't execute statement: $DBI::errstr";;	
+	    if ($prt) {print "\nrenamed station $station => $s_id\n";}
+	    $sth_history->execute( $s_id, 'renamed', $station_name{ lc $station } . " now $station" ) or warn "Can't execute statement: $DBI::errstr";;
 	    $station_name{ lc $station } = $station;
 	}
 	
@@ -442,7 +493,7 @@ sub getGlider {
 	    $sth_add->execute( $id ) or die "Can't execute statement: $DBI::errstr";;
 	    $s_id = $sth_add->{mysql_insertid};
 	    $glider_id{uc $id} = $s_id;
-	    print "\nnew glider $id => $s_id\n";
+	    if ($prt) {print "\nnew glider $id => $s_id\n";}
 	}
     }
 
@@ -520,7 +571,7 @@ sub handleAvailablity {
 	    lock( %gliders );
 	    foreach my $g ( keys %gliders ) {
 		if( scalar( keys %{$gliders{$g}} ) > 1 ) {
-		    print "$g:" . Dumper( $gliders{$g} ) . "\n";
+		    if ($prt) {print "$g:" . Dumper( $gliders{$g} ) . "\n";}
 		}
 	    }
 	    %gliders = ();
@@ -559,26 +610,26 @@ sub handleAvailablity {
 
 	my $now = time();
 	
-	print "\nup: ";
+	if ($prt) {print "\nup: ";}
 	my @missing;
 	foreach my $station ( sort keys %station_current ) {
 	    my $last = $station_current{$station};
 	    my $s_id = getStation( $sth_addstation, $sth_history, undef, $station );
 
-	    print "$station [". ($station_packets{$station}||0)."]";
+	    if ($prt) {print "$station [". ($station_packets{$station}||0)."]";}
 	    # if we didn't have a previous status it's a new station
 	    if( ! ($station_previous_check{$station}||0) ) {
 		$sth_log->execute( $s_id, $last, 'U' ) or die "Can't execute statement: $DBI::errstr";;
 		push( @missing, $station );
-		print " (logged UP)";
+		if ($prt) {print " (logged UP)";}
 	    }
-	    print ",";
+	    if ($prt) {print ",";}
 
 	    $sth->execute( $s_id, $last, 'U' ) or die "Can't execute statement: $DBI::errstr";;
 	    $sth_active->execute( $s_id ) or die "Can't execute statement: $DBI::errstr";;
 	}
 
-	print "done.\ndown: ";
+	if ($prt) {print "done.\ndown: ";}
 	foreach my $station ( sort keys %station_previous_check ) {
 	    my $last = $station_current{$station}||0;
 	    my $s_id = getStation( $sth_addstation, $sth_history, undef, $station );
@@ -587,18 +638,18 @@ sub handleAvailablity {
 	    # if it hasn't had a new time then we are down, 
 	    # possibly once or possibly multiple times
 	    if( ! $last ) {
-		print "$station";
+		if ($prt) {print "$station";}
 		if( $station_previous_check{$station} ) {
 		    $sth->execute( $s_id, $now, 'D' );
 		    $sth_log->execute( $s_id, $now, 'D' );
-		    print "  (logged down)";
+		    if ($prt) {print "  (logged down)";}
 		}
-		print ",";
+		if ($prt) {print ",";}
 	    }
 
 	    $station_previous_check{$station} = $last;
 	}
-	print "\n";
+	if ($prt) {print "\n";}
 	
 	foreach my $station (@missing) {
 	    $station_previous_check{$station} = $station_current{$station};
@@ -626,7 +677,7 @@ sub processStationLocationBeacon {
     if( ($stations_loc{$callsign}||'') ne "$lt/$lg" ) {
 	my $c = '';#getCountry($lt,$lg)
 	$st_station_loc->execute( $s_id, $lt, $lg, $altitude, $c );
-	print "station $callsign location update to $lt, $lg ($c)\n";
+	if ($prt) {print "station $callsign location update to $lt, $lg ($c)\n";}
 	$stations_loc{$callsign} = "$lt/$lg";
     }
 }
@@ -672,19 +723,19 @@ sub processStationDetailsBeacon {
     }
     
     if( ! $version ) {
-	print "****** [!version] $callsign: ".($comment||'')."\n";
+	if ($prt) {print "****** [!version] $callsign: ".($comment||'')."\n";}
     }
     
     {	
 	lock( %stations_ver );
 	if( ($stations_ver{$callsign}||'') ne "$version" ) {
 	    $st_station_ver->execute( $s_id, $version );
-	    print "station $callsign version update to ($version)\n";
+	    if ($prt) {print "station $callsign version update to ($version)\n";}
 	    $stations_ver{$callsign} = "$version";
 	}
     }
     
-    printf (":: %20s: ppm %0.1f/%0.1f db [%-70s]\n", $callsign,($ppm||-99),($dbadj||-99),$comment);
+    if ($prt) {printf (":: %20s: ppm %0.1f/%0.1f db [%-70s]\n", $callsign,($ppm||-99),($dbadj||-99),$comment);}
     
     if( defined($ppm) && defined($dbadj) )
     {
@@ -705,7 +756,7 @@ sub processStationDetailsBeacon {
 	if( ! $ppm && ! $dbadj ) {
 	    if( $cpu < 0.1 ) {
 		$stations_ppm{$callsign} ++;
-		print "$callsign: LOWCPU & noppm (flagged $stations_ppm{$callsign} times)". $comment.">>cpu $cpu $ppm ppm, $dbadj db \n";
+		if ($prt) {print "$callsign: LOWCPU & noppm (flagged $stations_ppm{$callsign} times)". $comment.">>cpu $cpu $ppm ppm, $dbadj db \n";}
 		
 		if( $stations_ppm{$callsign} > 10 ) { 
 		    $stations_status{$callsign} = 'N';
